requirements.txt
flask
flask_cors
pillow
imagehash
web3
python-dotenv
gunicorn